feat(python/bindings): remove request-reply API (#677)
# Description

Fixes #676.

Also:
 - refactor bindings by splitting them into modules.
