+++ data-plane/python/bindings/Taskfile.yaml
@@ -21,7 +21,7 @@ tasks:
     cmds:
       - uv run ruff check
       - uv run ruff format --check
-      - uv run mypy slim_bindings
+      - uv run mypy --ignore-missing-imports --explicit-package-bases .
 
   python:bindings:build:
     desc: "Build the Python bindings"
