+++ data-plane/python/bindings/tests/common.py
@@ -5,20 +5,20 @@ import slim_bindings
 
 
 async def create_svc(name: slim_bindings.PyName, secret):
-    provider = slim_bindings.PyIdentityProvider.SharedSecret(
+    provider = slim_bindings.PyIdentityProvider.SharedSecret(  # type: ignore
         identity=f"{name}", shared_secret=secret
     )
-    verifier = slim_bindings.PyIdentityVerifier.SharedSecret(
+    verifier = slim_bindings.PyIdentityVerifier.SharedSecret(  # type: ignore
         identity=f"{name}", shared_secret=secret
     )
     return await slim_bindings.create_pyservice(name, provider, verifier)
 
 
 async def create_slim(name: slim_bindings.PyName, secret):
-    provider = slim_bindings.PyIdentityProvider.SharedSecret(
+    provider = slim_bindings.PyIdentityProvider.SharedSecret(  # type: ignore
         identity=f"{name}", shared_secret=secret
     )
-    verifier = slim_bindings.PyIdentityVerifier.SharedSecret(
+    verifier = slim_bindings.PyIdentityVerifier.SharedSecret(  # type: ignore
         identity=f"{name}", shared_secret=secret
     )
     return await slim_bindings.Slim.new(name, provider, verifier)
